pyproject.toml
[project]
name = "telegram-mcp"
version = "0.1.0"
description = "Telegram MCP server for bidirectional LLM-human communication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
